v3/requirements.txt
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
et_xmlfile==2.0.0
idna==3.11
lxml==6.0.2
numpy==2.3.5
openpyxl==3.1.2
packaging==25.0
pandas==2.2.3
pdf2image==1.16.3
pdfminer.six==20221105
pdfplumber==0.10.3
pillow==12.0.0
pycparser==2.23
pypdfium2==5.1.0
pytesseract==0.3.10
python-dateutil==2.9.0.post0
python-docx==0.8.11
python-pptx==0.6.21
pytz==2025.2
requests==2.32.3
six==1.17.0
tzdata==2025.2
urllib3==2.5.0
xlsxwriter==3.2.9
